qemu-zcu102: fix nested tmux
If tmux is already active when we start then new windows & panes get the
environment of whoever started the session (or server?).

We were lucking out before when the tmux window & panes were inheriting
our env vars.  This now makes it very explicit and so works even if tmux
is already running.

* Move ENV vars that don't change to the top (global)
* save environment vars before doing main processing
* restore from env file if told to on the command line
* in tmux commands, tell subcmds to restore environment
* save the last envf file as a debug aid
  * It has the APU and PMU arguments we stopped printing

Signed-off-by: Bill Mills <bill.mills@linaro.org>