qemu-zcu102: attempt and fail to include vvfat support
QEMU supports creating a disk image from a directory of files.
This would be a nice feature to include.
HOWEVER, it does not appear to have been tested with the SD card device
type.

The first complaint was that SD cards could not be read-only.
(This is dumb as full size SD cards have a write protect slider.)
I added the :rw: option to get past this but it would be better not to.

The next and fatal (for now) hurdle was the currently packages Xilinx QEMU
requires SD cards to be powers of 2.  This is really dumb as real SD cards
are NEVER powers of 2.  I have complained about this before and Xilinx has
fixed it in later versions but Zephyr has not picked those up.

There may be more hurdles if we get past that one.  I don't know.

So I am keeping the code but disabling it.

Signed-off-by: Bill Mills <bill.mills@linaro.org>