qemu-zcu102: better temp cleanup for shell
Prior to this the shell was deleting the temp files while QEMU was still
running.

This kills all other panes in the current window before cleaning up.
This will include QEMU.
(unless the user is a tmux expert and broken it out to a new window.)

I don't know if kill-pane is sync or async but I added a small delay
"just in case".

The tmux-kill in this case could just be an exit,
but I kept it "just in case"

Signed-off-by: Bill Mills <bill.mills@linaro.org>