qemu-zcu102: fix make-image-targets when no helper present.
If there is no helper in the test base dir, then it should have done
nothing.  Instead it was calling make-image-targets anyway.

This was harmless as right now make-image-targets does nothing w/o the
helper but it is rather chatty about it.

Fixed the two typos that was causing the issue.
* missing -e caused it to always run, then
* missing "s" in targets caused it to never run

Signed-off-by: Bill Mills <bill.mills@linaro.org>