Fixup firmware images in my-extra-stuff for demo1 and demo2
my-extra-stuff can't really use symlinks so we have copies here.
make-target-helper could copy them in but I think it is more clear if the
user can look at the files there before they run.

We could copy them there when we build the image for docker but for now
we will manually update them.

Signed-off-by: Bill Mills <bill.mills@linaro.org>