test-sw/openamp: use existing modules in hacked.cpio.gz
Now that the base cpio file has modules that match the kernel we are using,
change make-image-targets not to delete them but only trim them.

As we may go back and forth on this make it easy to switch by refactoring
the callback functions and showing both options with one commented.

Signed-off-by: Bill Mills <bill.mills@linaro.org>