Demos: this is the boot log from the ST MP1 w/ generic-armv7a
The demo works but there are 1 to 3 kernel oops'es.
At least the one that always shows up is related to sound.
I think the sound modules are not available and the code does not handle
that well.  I suspect that all 3 oops'es are sound releated but I don't
know for sure.

Signed-off-by: Bill Mills <bill.mills@linaro.org>