make-image-targets: add extract support for deb and ipk
Allow deb and (Yocto Project) ipk files to be used as input to a
new-* command.

This does not handle OpenWRT ipk files as they have forked the format
away from being deb compatible.

None of the package format extractions handles the package info files or
script files.  These files are not included in the extraction.

Signed-off-by: Bill Mills <bill.mills@linaro.org>