demos: use stock dtb for stm32mp157*
With v6.6 and its new config we no longer have the audio boot time issue.
Therefore use the stock dtb built from v6.6.

The dtb should not be required to have a dash in it.
This was OK when we used -no-audio for the ST board but now without it
we need the simple ${BOARD}.dtb

Signed-off-by: Bill Mills <bill.mills@linaro.org>