docker: use only checked-in files for user-dev, qemu-zcu102, & demos
As it was, ignored and added files from a dirty working dir were put
into the docker image.

Instead use git to create an archive of each sub-dir and then extract
it in place.

As we are creating more tarfiles and markers, ignore them all.
zephyr-sdk tarfile is already checked in so it is ok but it is going away
soon anyway.

Ignore the working dirs xxx-temp-*

Signed-off-by: Bill Mills <bill.mills@linaro.org>