HACK: docker: force user "dev" to != 1000 for testing
This is a testing hack and should not be enabled when building
public container images.

To test when dev UID does not equal the users dev id, I do this.
I am always 1000 on my machine and the ubuntu user on ec2 is alwasy 1000
so this is the easiest way.

Signed-off-by: Bill Mills <bill.mills@linaro.org>