demos: in R5 combo demo, hide demo scripts that don't apply this boot
In the R5 combo demo (demo1-2-3) and the individual demos built from it
(demo1, demo2, and demo3), hide the demo scripts that don't apply to the
currently booted DTB configuration.

We already told the user which ones work but this is extra protection
for the curious.  The scripts can still be looked at in any boot as the
hidden scripts are just moved to the "hide" directory.

NOTE: ash does not do curly bracket globbing demo{1,2}* so we use the
more explicit syntax.

Signed-off-by: Bill Mills <bill.mills@linaro.org>